Veni, Vidi, Velocity
Changed all instances of acceleration to velocity in unit and hero.
Added a terminal velocity constant in unit, and appropriate checks to
enforce it.
Added a private keyword to the fields in the Rect class
Added an extra clause to the hero-x-movement, namely the case where no
direction is selected.
Added falling and falling checks to units.
Added a check so the hero can't jump in midair.