Portals - All the world is a stage
- People have their entrances and exits.

Portals (entrances and exits) are now placed by the generator.
Some modules had portals added.
Renamed several global variables in the .ino file to begin with a
underscore.
Changed the contains functions in Geo to handle pointers instead of
copies.
Changed the getPos in Rects to return a pointer instead of a copy.
Added the addUnit function to the .ino file, which adds the given unit
to the _units list and sets the position to the given tile.
Added entrance and exit tiles.
Renamed _archon to _hero in the .ino file.