Game Restart - Boot and Reboot was in a boat...
Added GameOver and HeroWon booleans, getters and setters in logic.
GamOver is true when the current map is finished, and HeroWon is true if
the hero won (made it to the exit) or false if not (died).
Added an innerloop after gamelogic but before drawlogic which determines
whether to put a new map on. If true, makes a new map, clears the unit
list and restarts the logic.
Fixed type0 modules being improperly recognized.
Fixed a module type including a portal tile, which would never be used.
Added a clearScene() function to SceneGenerator, which clears the entire
map.